Added theme-provider
Moved App.tsx to ./src
