+++ app/index.js
@@ -4,7 +4,9 @@
 
 import {AppRegistry} from 'react-native';
 
-import App from './App';
+import 'react-native-gesture-handler';
+import App from './src/App';
+
 import {name as appName} from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
